# Use netip.Addr instead of net.IP