--- a/conn_test.go
+++ b/conn_test.go
@@ -3,7 +3,7 @@ package conntrack_test
 import (
 	"fmt"
 	"log"
-	"net"
+	"net/netip"
 	"testing"
 
 	"github.com/mdlayher/netlink"
@@ -34,8 +34,8 @@ func ExampleConn_createUpdateFlow() {
 	// Set up a new Flow object using a given set of attributes.
 	f := conntrack.NewFlow(
 		17, 0,
-		net.ParseIP("2a00:1450:400e:804::200e"),
-		net.ParseIP("2a00:1450:400e:804::200f"),
+		netip.MustParseAddr("2a00:1450:400e:804::200e"),
+		netip.MustParseAddr("2a00:1450:400e:804::200f"),
 		1234, 80, 120, 0,
 	)
 
@@ -72,12 +72,12 @@ func ExampleConn_dumpFilter() {
 	}
 
 	f1 := conntrack.NewFlow(
-		6, 0, net.IPv4(1, 2, 3, 4), net.IPv4(5, 6, 7, 8),
+		6, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"),
 		1234, 80, 120, 0x00ff, // Set a connection mark
 	)
 
 	f2 := conntrack.NewFlow(
-		17, 0, net.ParseIP("2a00:1450:400e:804::200e"), net.ParseIP("2a00:1450:400e:804::200f"),
+		17, 0, netip.MustParseAddr("2a00:1450:400e:804::200e"), netip.MustParseAddr("2a00:1450:400e:804::200f"),
 		1234, 80, 120, 0xff00, // Set a connection mark
 	)
 
@@ -116,12 +116,12 @@ func ExampleConn_flushFilter() {
 	}
 
 	f1 := conntrack.NewFlow(
-		6, 0, net.IPv4(1, 2, 3, 4), net.IPv4(5, 6, 7, 8),
+		6, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"),
 		1234, 80, 120, 0x00ff, // Set a connection mark
 	)
 
 	f2 := conntrack.NewFlow(
-		17, 0, net.ParseIP("2a00:1450:400e:804::200e"), net.ParseIP("2a00:1450:400e:804::200f"),
+		17, 0, netip.MustParseAddr("2a00:1450:400e:804::200e"), netip.MustParseAddr("2a00:1450:400e:804::200f"),
 		1234, 80, 120, 0xff00, // Set a connection mark
 	)
 
@@ -155,7 +155,7 @@ func ExampleConn_delete() {
 	}
 
 	f := conntrack.NewFlow(
-		6, 0, net.IPv4(1, 2, 3, 4), net.IPv4(5, 6, 7, 8),
+		6, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"),
 		1234, 80, 120, 0,
 	)