--- a/expect_integration_test.go
+++ b/expect_integration_test.go
@@ -3,7 +3,7 @@
 package conntrack
 
 import (
-	"net"
+	"net/netip"
 	"testing"
 
 	"golang.org/x/sys/unix"
@@ -27,7 +27,7 @@ func TestConnCreateExpect(t *testing.T) {
 	c, _, err := makeNSConn()
 	require.NoError(t, err)
 
-	f := NewFlow(6, 0, net.IPv4(1, 2, 3, 4), net.IPv4(5, 6, 7, 8), 42000, 21, 120, 0)
+	f := NewFlow(6, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"), 42000, 21, 120, 0)
 
 	err = c.Create(f)
 	require.NoError(t, err, "unexpected error creating flow", f)
@@ -37,8 +37,8 @@ func TestConnCreateExpect(t *testing.T) {
 		TupleMaster: f.TupleOrig,
 		Tuple: Tuple{
 			IP: IPTuple{
-				SourceAddress:      net.IPv4(1, 2, 3, 4),
-				DestinationAddress: net.IPv4(5, 6, 7, 8),
+				SourceAddress:      netip.MustParseAddr("1.2.3.4"),
+				DestinationAddress: netip.MustParseAddr("5.6.7.8"),
 			},
 			Proto: ProtoTuple{
 				Protocol:        6,
@@ -48,8 +48,8 @@ func TestConnCreateExpect(t *testing.T) {
 		},
 		Mask: Tuple{
 			IP: IPTuple{
-				SourceAddress:      net.IPv4(255, 255, 255, 255),
-				DestinationAddress: net.IPv4(255, 255, 255, 255),
+				SourceAddress:      netip.MustParseAddr("255.255.255.255"),
+				DestinationAddress: netip.MustParseAddr("255.255.255.255"),
 			},
 			Proto: ProtoTuple{
 				Protocol:        6,