--- a/flow_integration_test.go
+++ b/flow_integration_test.go
@@ -3,7 +3,7 @@
 package conntrack
 
 import (
-	"net"
+	"net/netip"
 	"testing"
 
 	"golang.org/x/sys/unix"
@@ -35,7 +35,7 @@ func TestConnCreateFlows(t *testing.T) {
 
 	// Create IPv4 flows
 	for i := 1; i <= numFlows; i++ {
-		f = NewFlow(6, 0, net.IPv4(1, 2, 3, 4), net.IPv4(5, 6, 7, 8), 1234, uint16(i), 120, 0)
+		f = NewFlow(6, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"), 1234, uint16(i), 120, 0)
 
 		err = c.Create(f)
 		require.NoError(t, err, "creating IPv4 flow", i)
@@ -45,8 +45,8 @@ func TestConnCreateFlows(t *testing.T) {
 	for i := 1; i <= numFlows; i++ {
 		err = c.Create(NewFlow(
 			17, 0,
-			net.ParseIP("2a00:1450:400e:804::200e"),
-			net.ParseIP("2a00:1450:400e:804::200f"),
+			netip.MustParseAddr("2a00:1450:400e:804::200e"),
+			netip.MustParseAddr("2a00:1450:400e:804::200f"),
 			1234, uint16(i), 120, 0,
 		))
 		require.NoError(t, err, "creating IPv6 flow", i)
@@ -81,8 +81,8 @@ func TestConnFlush(t *testing.T) {
 	// Create IPv4 flow
 	err = c.Create(NewFlow(
 		6, 0,
-		net.IPv4(1, 2, 3, 4),
-		net.IPv4(5, 6, 7, 8),
+		netip.MustParseAddr("1.2.3.4"),
+		netip.MustParseAddr("5.6.7.8"),
 		1234, 80, 120, 0,
 	))
 	require.NoError(t, err, "creating IPv4 flow")
@@ -90,8 +90,8 @@ func TestConnFlush(t *testing.T) {
 	// Create IPv6 flow
 	err = c.Create(NewFlow(
 		17, 0,
-		net.ParseIP("2a00:1450:400e:804::200e"),
-		net.ParseIP("2a00:1450:400e:804::200f"),
+		netip.MustParseAddr("2a00:1450:400e:804::200e"),
+		netip.MustParseAddr("2a00:1450:400e:804::200f"),
 		1234, 80, 120, 0,
 	))
 	require.NoError(t, err, "creating IPv6 flow")
@@ -130,8 +130,8 @@ func TestConnFlushFilter(t *testing.T) {
 	// Create IPv4 flow
 	err = c.Create(NewFlow(
 		6, 0,
-		net.IPv4(1, 2, 3, 4),
-		net.IPv4(5, 6, 7, 8),
+		netip.MustParseAddr("1.2.3.4"),
+		netip.MustParseAddr("5.6.7.8"),
 		1234, 80, 120, 0,
 	))
 	require.NoError(t, err, "creating IPv4 flow")
@@ -139,8 +139,8 @@ func TestConnFlushFilter(t *testing.T) {
 	// Create IPv6 flow with mark
 	err = c.Create(NewFlow(
 		17, 0,
-		net.ParseIP("2a00:1450:400e:804::200e"),
-		net.ParseIP("2a00:1450:400e:804::200f"),
+		netip.MustParseAddr("2a00:1450:400e:804::200e"),
+		netip.MustParseAddr("2a00:1450:400e:804::200f"),
 		1234, 80, 120, 0xff00,
 	))
 	require.NoError(t, err, "creating IPv6 flow")
@@ -174,8 +174,8 @@ func TestConnCreateDeleteFlows(t *testing.T) {
 	for i := 1; i <= numFlows; i++ {
 		f = NewFlow(
 			17, 0,
-			net.ParseIP("2a00:1450:400e:804::223e"),
-			net.ParseIP("2a00:1450:400e:804::223f"),
+			netip.MustParseAddr("2a00:1450:400e:804::223e"),
+			netip.MustParseAddr("2a00:1450:400e:804::223f"),
 			1234, uint16(i), 120, 0,
 		)
 
@@ -199,8 +199,8 @@ func TestConnCreateUpdateFlow(t *testing.T) {
 
 	f := NewFlow(
 		17, 0,
-		net.ParseIP("1.2.3.4"),
-		net.ParseIP("5.6.7.8"),
+		netip.MustParseAddr("1.2.3.4"),
+		netip.MustParseAddr("5.6.7.8"),
 		1234, 5678, 120, 0,
 	)
 
@@ -262,8 +262,8 @@ func TestConnUpdateError(t *testing.T) {
 
 	f := NewFlow(
 		17, 0,
-		net.ParseIP("1.2.3.4"),
-		net.ParseIP("5.6.7.8"),
+		netip.MustParseAddr("1.2.3.4"),
+		netip.MustParseAddr("5.6.7.8"),
 		1234, 5678, 120, 0,
 	)
 
@@ -280,10 +280,10 @@ func TestConnCreateGetFlow(t *testing.T) {
 	require.NoError(t, err)
 
 	flows := map[string]Flow{
-		"v4m1": NewFlow(17, 0, net.ParseIP("1.2.3.4"), net.ParseIP("5.6.7.8"), 1234, 5678, 120, 0),
-		"v4m2": NewFlow(17, 0, net.ParseIP("10.0.0.1"), net.ParseIP("10.0.0.2"), 24000, 80, 120, 0),
-		"v6m1": NewFlow(17, 0, net.ParseIP("2a12:1234:200f:600::200a"), net.ParseIP("2a12:1234:200f:600::200b"), 6554, 53, 120, 0),
-		"v6m2": NewFlow(17, 0, net.ParseIP("900d:f00d:24::7"), net.ParseIP("baad:beef:b00::b00"), 1323, 22, 120, 0),
+		"v4m1": NewFlow(17, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"), 1234, 5678, 120, 0),
+		"v4m2": NewFlow(17, 0, netip.MustParseAddr("10.0.0.1"), netip.MustParseAddr("10.0.0.2"), 24000, 80, 120, 0),
+		"v6m1": NewFlow(17, 0, netip.MustParseAddr("2a12:1234:200f:600::200a"), netip.MustParseAddr("2a12:1234:200f:600::200b"), 6554, 53, 120, 0),
+		"v6m2": NewFlow(17, 0, netip.MustParseAddr("900d:f00d:24::7"), netip.MustParseAddr("baad:beef:b00::b00"), 1323, 22, 120, 0),
 	}
 
 	for n, f := range flows {
@@ -306,7 +306,7 @@ func TestDumpZero(t *testing.T) {
 	c, _, err := makeNSConn()
 	require.NoError(t, err)
 
-	f := NewFlow(17, 0, net.ParseIP("1.2.3.4"), net.ParseIP("5.6.7.8"), 1234, 5678, 120, 0xff000000)
+	f := NewFlow(17, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"), 1234, 5678, 120, 0xff000000)
 
 	f.CountersOrig.Bytes = 1337
 	f.CountersReply.Bytes = 9001
@@ -332,10 +332,10 @@ func TestConnDumpFilter(t *testing.T) {
 	require.NoError(t, err)
 
 	flows := map[string]Flow{
-		"v4m1": NewFlow(17, 0, net.ParseIP("1.2.3.4"), net.ParseIP("5.6.7.8"), 1234, 5678, 120, 0xff000000),
-		"v4m2": NewFlow(17, 0, net.ParseIP("10.0.0.1"), net.ParseIP("10.0.0.2"), 24000, 80, 120, 0x00ff0000),
-		"v6m1": NewFlow(17, 0, net.ParseIP("2a12:1234:200f:600::200a"), net.ParseIP("2a12:1234:200f:600::200b"), 6554, 53, 120, 0x0000ff00),
-		"v6m2": NewFlow(17, 0, net.ParseIP("900d:f00d:24::7"), net.ParseIP("baad:beef:b00::b00"), 1323, 22, 120, 0x000000ff),
+		"v4m1": NewFlow(17, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"), 1234, 5678, 120, 0xff000000),
+		"v4m2": NewFlow(17, 0, netip.MustParseAddr("10.0.0.1"), netip.MustParseAddr("10.0.0.2"), 24000, 80, 120, 0x00ff0000),
+		"v6m1": NewFlow(17, 0, netip.MustParseAddr("2a12:1234:200f:600::200a"), netip.MustParseAddr("2a12:1234:200f:600::200b"), 6554, 53, 120, 0x0000ff00),
+		"v6m2": NewFlow(17, 0, netip.MustParseAddr("900d:f00d:24::7"), netip.MustParseAddr("baad:beef:b00::b00"), 1323, 22, 120, 0x000000ff),
 	}
 
 	// Expect empty result from empty table dump
@@ -372,7 +372,7 @@ func BenchmarkCreateDeleteFlow(b *testing.B) {
 		b.Fatal(err)
 	}
 
-	f := NewFlow(6, 0, net.IPv4(1, 2, 3, 4), net.IPv4(5, 6, 7, 8), 1234, 80, 120, 0)
+	f := NewFlow(6, 0, netip.MustParseAddr("1.2.3.4"), netip.MustParseAddr("5.6.7.8"), 1234, 80, 120, 0)
 
 	for n := 0; n < b.N; n++ {
 		err = c.Create(f)