--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,6 @@ module github.com/ti-mo/conntrack
 go 1.20
 
 require (
-	github.com/google/go-cmp v0.6.0
 	github.com/mdlayher/netlink v1.7.2
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.8.4
@@ -14,6 +13,7 @@ require (
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/google/go-cmp v0.6.0 // indirect
 	github.com/josharian/native v1.1.0 // indirect
 	github.com/mdlayher/socket v0.4.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect