@antoninbas 2023-10-10
Before the change:
```
$ make bench-integration 
sudo modprobe -a nf_nat nf_conntrack xt_conntrack xt_MASQUERADE nf_conntrack_netlink
go test -bench=. -tags=integration -exec sudo ./...
goos: linux
goarch: amd64
pkg: github.com/ti-mo/conntrack
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkExpectUnmarshal-4            	156274430	         7.752 ns/op	       0 B/op	       0 allocs/op
BenchmarkCreateDeleteFlow-4           	   34616	     34257 ns/op	   20368 B/op	     180 allocs/op
BenchmarkFlowUnmarshal-4              	  242976	      4404 ns/op	    2952 B/op	      99 allocs/op
BenchmarkStatusUnmarshalAttribute-4   	23452940	        49.31 ns/op
BenchmarkIPTupleUnmarshal/ipv4-4      	 1959086	       619.5 ns/op	      48 B/op	       6 allocs/op
BenchmarkIPTupleUnmarshal/ipv6-4      	 1991146	       595.6 ns/op	      64 B/op	       4 allocs/op
PASS
ok  	github.com/ti-mo/conntrack	10.908s

$ go test -v -run=TestFlowsMemoryUsage
=== RUN   TestFlowsMemoryUsage
    flow_test.go:560: Num flows: 100000
    flow_test.go:561: Live objects: 1100511
    flow_test.go:562: Bytes of allocated heap objects: 67448680
    flow_test.go:563: Avg bytes per flow: 674
--- PASS: TestFlowsMemoryUsage (1.05s)
PASS
ok  	github.com/ti-mo/conntrack	1.064s
```

After the change:
```
$ make bench-integration 
sudo modprobe -a nf_nat nf_conntrack xt_conntrack xt_MASQUERADE nf_conntrack_netlink
go test -bench=. -tags=integration -exec sudo ./...
goos: linux
goarch: amd64
pkg: github.com/ti-mo/conntrack
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkExpectUnmarshal-4            	155513721	         7.733 ns/op	       0 B/op	       0 allocs/op
BenchmarkCreateDeleteFlow-4           	   35670	     32901 ns/op	   20400 B/op	     188 allocs/op
BenchmarkFlowUnmarshal-4              	  262989	      4239 ns/op	    2856 B/op	      93 allocs/op
BenchmarkStatusUnmarshalAttribute-4   	23563472	        49.02 ns/op
BenchmarkIPTupleUnmarshal/ipv4-4      	 2133122	       550.4 ns/op	      16 B/op	       4 allocs/op
BenchmarkIPTupleUnmarshal/ipv6-4      	 2001910	       598.8 ns/op	      64 B/op	       4 allocs/op
PASS
ok  	github.com/ti-mo/conntrack	10.770s

$ go test -v -run=TestFlowsMemoryUsage
=== RUN   TestFlowsMemoryUsage
    flow_test.go:558: Num flows: 100000
    flow_test.go:559: Live objects: 500503
    flow_test.go:560: Bytes of allocated heap objects: 57847064
    flow_test.go:561: Avg bytes per flow: 578
--- PASS: TestFlowsMemoryUsage (0.97s)
PASS
ok  	github.com/ti-mo/conntrack	0.984s
```
 
Some observations:
* memory usage for a `Flow` object goes down from 674B to 578B
* `BenchmarkCreateDeleteFlow` shows slightly increased memory usage. This is because the benchmark keeps marshaling the same flow and the marshaling process is able to reuse the `net.IP` slices from the `Flow` object, without having to allocate new ones. With `netip.Addr`, the marshaling / serialization requires creating byte slices. However, in practice: 1) actual code is unlikely to marshal the _same_ flow over and over again, 2) I believe that most users of the library would be interested in reading conntrack entries, hence unmarshaling flows.