@antoninbas 2023-10-11
@ti-mo thanks for the quick review. I will address your comments today.

Note that I think it may be a good idea to have one minor release (v0.5.0) before merging this change, and one minor release (v0.6.0) just after merging this change. v0.5.0 would be the last release to support `net.IP`, and it gives consumers of this library the opportunity to consume all the latest fixes / improvements that have been merged into the master branch, in a non-disruptive way. From v0.6.0 onwards, the API is changed (`net.IP` -> `netip.Addr`) and consumers have to update their code. The v0.5.0 and v0.6.0 release notes in Github should include the appropriate disclaimers. What do you think?