@ti-mo 2023-10-11
Yep, was considering that as well. I'll bump the Go version and do a few more dependency updates, then release 0.5.0.