@antoninbas 2023-10-12
@ti-mo I have addressed your comments, save for the one about `netlink.AttributeDecoder.Reset()`. I believe `AttributeDecoder` should not be re-used as a rule, and we make an exception for the benchmark. Even if you feel strongly about it, it seems orthogonal to this PR, and something that can be taken care of as a follow-up, especially since it involves another library.