@ti-mo 2023-10-16
I gave this a rebase and a `go mod tidy` as go-cmp was removed. I've just released 0.4.1 and I'll release 0.5.0 when this PR is merged.